MAT-8269:  Using overrides to avoid a huge if..then..else class in MeasureService
