+++ src/test/java/gov/cms/madie/models/cqm/datacriteria/basetypes/DataElementTest.java
@@ -8,6 +8,8 @@ import java.time.ZoneId;
 import java.time.ZonedDateTime;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrowsExactly;
 
 class DataElementTest {
 
@@ -27,5 +29,18 @@ class DataElementTest {
     encounterPerformed.setAuthorDatetime(dateTime);
     ZonedDateTime shiftedDateTime = encounterPerformed.shiftDateByYear(dateTime, -100000);
     assertThat(shiftedDateTime).isEqualTo(dateTime.withYear(1900));
+  }
+
+  @Test
+  void testSuperClassThrowsException() {
+    DataElement dataElement = new DataElement();
+
+    ShiftDatesException thrown =
+        assertThrowsExactly(
+            ShiftDatesException.class,
+            () -> dataElement.shiftDates(0),
+            "Expected doThing() to throw, but it didn't");
+
+    assertEquals(thrown.getMessage(), "There isn't an implementation of ShiftDates for this type");
   }
 }
